method formingJsonForAnswer and method findBiggestSum were added

+++ src/main/java/BoxOfThings/Cloth.java
@@ -1,7 +1,7 @@
 package BoxOfThings;
 
-public class Cloth extends Things{
-    public Cloth(String title, String date, long sum) {
+public class Cloth extends Things {
+    public Cloth(String title, String date, int sum) {
         super(title, date, sum);
     }
 }
