+++ src/main/java/BoxOfThings/Finances.java
@@ -1,7 +1,7 @@
 package BoxOfThings;
 
-public class Finances extends Things{
-    public Finances(String title, String date, long sum) {
+public class Finances extends Things {
+    public Finances(String title, String date, int sum) {
         super(title, date, sum);
     }
 }
