+++ src/main/java/BoxOfThings/Life.java
@@ -1,7 +1,7 @@
 package BoxOfThings;
 
-public class Life extends Things{
-    public Life(String title, String date, long sum) {
+public class Life extends Things {
+    public Life(String title, String date, int sum) {
         super(title, date, sum);
     }
 }
