+++ src/main/java/BoxOfThings/Other.java
@@ -1,7 +1,7 @@
 package BoxOfThings;
 
-public class Other extends Things{
-    public Other(String title, String date, long sum) {
+public class Other extends Things {
+    public Other(String title, String date, int sum) {
         super(title, date, sum);
     }
 }
